Code/docker/requirements.txt
apache-airflow
mlflow
pandas
PyYAML
openpyxl
scikit-learn
joblib
optuna[visualization]
xgboost
lightgbm
catboost
seaborn
matplotlib
sqlalchemy
psycopg2-binary
apache-airflow
kafka-python
requests